requirements.txt
# Voice Assistant Dependencies

# Audio Processing
pyaudio>=0.2.11
numpy>=1.24.0

# STT - Faster Whisper
faster-whisper>=0.10.0

# Wake Word Detection
openwakeword>=0.5.0

# TTS - Piper
piper-tts>=1.0.0

# HTTP Requests
requests>=2.31.0

# Optional: Sound playback for notifications
pygame>=2.5.0

# Optional: Ollama client (if using Ollama API directly)
# ollama>=0.1.0

# OS compatibility
importlib-metadata>=6.0.0; python_version < "3.10"


